Add files via upload
The codes represent a computer vision model that classifies with high accuracy 10 different types of conditions from chest x-rays. I have used a subset of a medical image dataset and these are often trained on millions of images, it is unrealistic for me to achieve an extremely high accuracy. Moreover, it is important to note that binary versus multiclass classification yields very different accuracies in CV models. In the same experiment that achieved above 80% classification for binary classification, the accuracy dropped to 65% accuracy when 3 classes were introduced. With that being said, I am aiming for an accuracy of between 50 and 60% for my CV model.